pyproject.toml
[tool.poetry]
name = "zpl-image-extractor"
version = "0.1.0"
description = ""
authors = ["pyduh <eduardo.pereira@olist.com>"]
readme = "README.md"
packages = [{include = "zpl_image_extractor"}]

[tool.poetry.dependencies]
python = "~3.11"
Pillow = "^9.4.0"
click = "^8.1.3"
pyinstaller = "^5.7.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
